Create Advanced Code
This code can save the images according to the name given by the user. for example : 20 images names as plants uploaded we can divide the images into 2 group like index 0 to 9 as picture and index 10 to 19 as image. The result will be picture 1 to picture 10 and image 1 to image 10 in the same zip file.